IAR tool - all logic should be within IAR class
Exporter got too smart, and there is a logic also in flasher,
builder. These should be unified within one IAR. I'll rename it
to IAR as it supports different toolchains (needs testing and
verification).
