+++ project_generator/tools/iar_definitions.py
